requirements.txt
Flask
gunicorn
numpy
pillow
joblib
opencv-python-headless
requests
